+++ app/Http/Controllers/Admin/HomeController.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Http\Controllers\Admin;
+
+use App\Http\Controllers\Controller;
+
+use Illuminate\Http\Request;
+
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
+
+use App\Models\CustomUserTable;
+use App\Models\Ticket;
+use App\Models\Office;
+
+class HomeController extends Controller
+{
+
+}